third_party/WebKit/LayoutTests/geolocation-api/reentrant-error.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/geolocation-mock.js"></script>
</head>
<body>
<script>
description("Tests that reentrant calls to Geolocation methods from the error callback are OK.");

var mockMessage = 'test';

var error;

geolocationMock.then(mock => {
    mock.setGeolocationPermission(true);
    mock.setGeolocationPositionUnavailableError(mockMessage);

    var errorCallbackInvoked = false;
    navigator.geolocation.getCurrentPosition(function(p) {
        testFailed('Success callback invoked unexpectedly');
        finishJSTest();
    }, function(e) {
        if (errorCallbackInvoked) {
            testFailed('Error callback invoked unexpectedly');
            finishJSTest();
        }
        errorCallbackInvoked = true;

        error = e;
        shouldBe('error.code', 'error.POSITION_UNAVAILABLE');
        shouldBe('error.message', 'mockMessage');
        debug('');
        continueTest();
    });

    function continueTest() {
        mockMessage += ' repeat';

        mock.setGeolocationPositionUnavailableError(mockMessage);

        navigator.geolocation.getCurrentPosition(function(p) {
            testFailed('Success callback invoked unexpectedly');
            finishJSTest();
        }, function(e) {
            error = e;
            shouldBe('error.code', 'error.POSITION_UNAVAILABLE');
            shouldBe('error.message', 'mockMessage');
            finishJSTest();
        });
    }
});

window.jsTestIsAsync = true;
</script>
</body>
</html>
